Redefined how FXP_max_ld gets initialized: now from first invalid pos integer - half tiniest; if any value is >= that, then it's +INF. The previous method (from largest valid value, and checking if any value was > that, then +INF) was causing a corner case problem for 31 frac bits when running get_pow2_target(-tiniest) which was returning +INF instead of a value smaller than but very close to 1.
exp_l() and pow10_l() implemented using pow2_l()

Fixed a bug in mul_distrib_l(): xbyb was being used before
initialization.

Simplified and unified the tuple variable names on fxp.c
and fxp_l.c

+++ print_as_bits.c
@@ -40,7 +40,7 @@ void print_int_as_bin(int n, int width)
 
 void print_long_as_bin(long n)
 {
-        int an;
+        unsigned long an;
         if (n < 0) {
                 an = -n;
                 printf("-");
@@ -49,7 +49,7 @@ void print_long_as_bin(long n)
         }
         int i = LONG_BITS;
         while (i > 0) {
-                int bit = (an >> (i - 1)) & 1;
+                int bit = (int) ((an >> (i - 1)) & 1ul);
                 printf("%d", bit);
                 i--;
         }
@@ -70,7 +70,7 @@ void print_ulong_as_bin(unsigned long n)
 {
         int i = LONG_BITS;
         while (i > 0) {
-                int bit = (n >> (i - 1)) & 1;
+                int bit = (n >> (i - 1)) & 1ul;
                 printf("%d", bit);
                 i--;
         }
